bug of create dir.

+++ helper.js
@@ -13,15 +13,18 @@
   getOutWindow: function(){return GuiNode.Window.get()},
   saveWindowSize: function(){
     var fs=require('fs');
-    var pkg=JSON.parse(fs.readFileSync('package.json'));
-    var h=document.documentElement.clientHeight;
-    var w=document.documentElement.clientWidth;
-    if((h!=pkg.window.height || w!=pkg.window.width) && 
-      (w>100 && h>100))
+    if(fs.existsSync('package.json'))
     {
-      pkg.window.height=Math.floor(h);
-      pkg.window.width=Math.floor(w);
-      fs.writeFileSync('package.json',JSON.stringify(pkg));
+      var pkg=JSON.parse(fs.readFileSync('package.json'));
+      var h=document.documentElement.clientHeight;
+      var w=document.documentElement.clientWidth;
+      if((h!=pkg.window.height || w!=pkg.window.width) && 
+        (w>100 && h>100))
+      {
+        pkg.window.height=Math.floor(h);
+        pkg.window.width=Math.floor(w);
+        fs.writeFileSync('package.json',JSON.stringify(pkg));
+      }
     }
   },
   }}
